requirements.txt
certifi==2020.12.5
chardet==4.0.0
Click==7.0
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==19.9.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-dotenv==0.17.0
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
tmdbv3api==1.7.5
urllib3==1.26.4
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
